#1434 cleanup
