+++ vuu/src/test/scala/org/finos/vuu/wsapi/helpers/TestExtension.scala
@@ -8,7 +8,6 @@ import org.finos.vuu.core.module.{ModuleFactory, ModuleFactoryNode, TableDefCont
 import org.finos.vuu.core.table.{DataTable, TableContainer}
 import org.finos.vuu.provider.{MockProvider, Provider, ProviderContainer}
 
-
 object TestExtension {
   implicit class ModuleFactoryExtension(val moduleFactoryNode: ModuleFactoryNode) {
 
@@ -32,36 +31,4 @@ object TestExtension {
     }
 
   }
-}
-class TestModuleBuilder(moduleName:String)(implicit clock: Clock, lifecycle: LifecycleContainer, tableDefContainer: TableDefContainer){
-
-  private var moduleFactory = ModuleFactory.withNamespace(moduleName)
-
-  def withTable(tableDef: TableDef, viewPortDef: ViewPortDef): TestModuleBuilder = {
-    moduleFactory = moduleFactory.addTable(
-      tableDef,
-      (table, _) => new MockProvider(table),
-      (_, _, _, _) => viewPortDef
-    )
-    this
-  }
-
-  def build(moduleName: String, tableDef: TableDef, viewPortDef: ViewPortDef)(implicit clock: Clock, lifecycle: LifecycleContainer, tableDefContainer: TableDefContainer): ViewServerModule =
-  ModuleFactory.withNamespace(moduleName)
-      .addTable(
-        tableDef,
-        (table, _) => new MockProvider(table),
-        (_, _, _, _) => viewPortDef
-      )
-      .asModule()
-
-  def build(moduleName: String, tableDef: TableDef)(implicit clock: Clock, lifecycle: LifecycleContainer, tableDefContainer: TableDefContainer): ViewServerModule =
-    ModuleFactory.withNamespace(moduleName)
-      .addTable(
-        tableDef,
-        (table, _) => new MockProvider(table)
-      )
-      .asModule()
-
-
 }
