+++ vuu/src/test/scala/org/finos/vuu/wsapi/helpers/TestProvider.scala
@@ -29,11 +29,3 @@ class TestProvider(table: DataTable, fakeDataSource: FakeDataSource)(implicit cl
 
   override val lifecycleId: String = s"TestProvider ${table.name}"
 }
-class FakeDataSource(rows: Map[String, Map[String, Any]]) {
-  type RowKey = String
-  type ColumnName = String
-
-  def get(): Map[RowKey, Map[ColumnName, Any]] = {
-    rows
-  }
-}
